Add faye monkeypatch to fix ssl
Slack updated the cert for their API and this broke the bot. Seems like
this new Let's Encrypt cert they're using could not be verified by
the existing ad hoc cert verification code in faye, not sure why. But
as you can see in the comments in the monkeypatch file, other people
have run into this problem and I was able to grab their patch, seems
to work.